requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
chromadb==1.5.0
sentence-transformers==5.2.2
requests==2.31.0
beautifulsoup4==4.14.3
python-dotenv==1.0.0
pydantic==2.12.5
pytest==9.0.2
httpx==0.28.1
